Add files via upload
## Disclaimer

This tool is intended for ethical penetration testing and security research purposes only. It is strictly prohibited to use this tool for any illegal or malicious activities. You are solely responsible for any use of this tool. The developer is not liable for any misuse or illegal use of this tool.

You must obtain explicit permission from the owner before using this tool on any system or network. Unauthorized use of this tool is a violation of the law and may result in legal consequences.